+++ lib/core/test/test_dice.cpp
@@ -21,7 +21,7 @@ TEST(DiceTest, generate_result)
 TEST(DiceTest, count_success)
 {
    dice::D8 sequence(10);
-   size_t i = 0;
+   uint32_t i = 0;
    for (auto & val : sequence) {
       val(i++);
    }
